Refactor: 댓글 수정/삭제 API 위치 변경
