+++ web/components/Comment/CommentItem/CommentItem.tsx
@@ -1,16 +1,47 @@
-import { MouseEventHandler } from 'react';
-import { Comment } from '../../../types/comment';
+import { Comment, CreateOrUpdateComment } from '../../../types/comment';
+import { deleteComment, updateComment } from '../../../apis/CommentAPI';
 import Profile from '../../Profile';
 import * as S from './CommentItem.style';
 
 interface Props {
   comment: Comment;
-  onClickUpdate: MouseEventHandler;
-  onClickDelete: MouseEventHandler;
+  folderId: number;
 }
 
-const CommentItem = ({ comment, onClickUpdate, onClickDelete }: Props) => {
+const CommentItem = ({ comment, folderId }: Props) => {
   const { id, content, user, createdAt } = comment;
+  const tempUserID: number = 5;
+  const tempToken: string =
+    'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJyb2xlcyI6WyJST0xFX1VTRVIiXSwiaXNzIjoicHJncm1zIiwiZXhwIjoxNjYwMTI4ODYxLCJpYXQiOjE2NjAxMjUyNjEsImVtYWlsIjoianVuZ21pbWluZ0BnbWFpbC5jb20ifQ.F5N76kkVG2WGgL-A5cLQi7cpSClfpA1CPqIEMNHCh3u9CiRXFy00pKzEpxaeIkVMLn-L1MrJ0drDC5nttAWtsw';
+
+  const handleClickUpdateComment = () => {
+    const newComment: CreateOrUpdateComment = {
+      id,
+      content,
+      folderId,
+      userId: tempUserID,
+    };
+    console.log(newComment);
+
+    try {
+      const res = updateComment(newComment, tempToken);
+      console.log(res);
+    } catch (error) {
+      console.log(error);
+    }
+  };
+
+  const handleClickDeleteComment = () => {
+    const confirmDelete = confirm('정말 삭제하시겠습니까?');
+    if (!confirmDelete) return;
+
+    try {
+      const res = deleteComment(id, tempToken);
+      console.log(res);
+    } catch (error) {
+      console.log(error);
+    }
+  };
   return (
     <S.Container key={id}>
       <S.HeaderWrapper>
@@ -22,8 +53,8 @@ const CommentItem = ({ comment, onClickUpdate, onClickDelete }: Props) => {
         />
         {5 === user.id && (
           <S.ButtonsWrapper>
-            <S.Button onClick={onClickUpdate}>수정</S.Button>|
-            <S.Button onClick={onClickDelete}>삭제</S.Button>
+            <S.Button onClick={handleClickUpdateComment}>수정</S.Button>|
+            <S.Button onClick={handleClickDeleteComment}>삭제</S.Button>
           </S.ButtonsWrapper>
         )}
       </S.HeaderWrapper>
